Split CircleTool, LineTool and RangeSliderTool
